Almost Finished With Script
The script for the Tic Tac Toe game is almost finished apart from a few
components. The player can no longer click in a box with an 'O' in it.
Also, the computer's choice has been refined. Finally, I started writing
code in order to check for a game win.